+++ src/test/java/chess/domain/board/BoardTest.java
@@ -411,53 +411,6 @@ public class BoardTest {
     }
 
     @Test
-    void 같은_진영의_기물_가치_합을_구할_수_있다() {
-        /*
-
-        .KR.....
-        P.PB....
-        .P..Q...
-        ........
-        .....nq.
-        .....p.p
-        .....pp.
-        ....rk..
-
-         */
-
-        // given
-        boardMap.put(Position.of(E, ONE), new Rook(WHITE));
-        boardMap.put(Position.of(F, ONE), new King(WHITE));
-        boardMap.put(Position.of(F, TWO), new Pawn(WHITE));
-        boardMap.put(Position.of(G, TWO), new Pawn(WHITE));
-        boardMap.put(Position.of(F, THREE), new Pawn(WHITE));
-        boardMap.put(Position.of(H, THREE), new Pawn(WHITE));
-        boardMap.put(Position.of(F, FOUR), new Knight(WHITE));
-        boardMap.put(Position.of(G, FOUR), new Queen(WHITE));
-
-        boardMap.put(Position.of(A, SEVEN), new Pawn(BLACK));
-        boardMap.put(Position.of(B, SIX), new Pawn(BLACK));
-        boardMap.put(Position.of(B, EIGHT), new King(BLACK));
-        boardMap.put(Position.of(C, SEVEN), new Pawn(BLACK));
-        boardMap.put(Position.of(C, EIGHT), new Rook(BLACK));
-        boardMap.put(Position.of(D, SEVEN), new Bishop(BLACK));
-        boardMap.put(Position.of(E, SIX), new Queen(BLACK));
-
-        final Board board = new Board(boardMap);
-
-        // when
-        final double priceByWhite = board.calculatePriceBySide(WHITE);
-        final double priceByBlack = board.calculatePriceBySide(BLACK);
-
-        // then
-        final SoftAssertions softAssertions = new SoftAssertions();
-        softAssertions.assertThat(priceByWhite).isEqualTo(19.5);
-        softAssertions.assertThat(priceByBlack).isEqualTo(20.0);
-
-        softAssertions.assertAll();
-    }
-
-    @Test
     void 킹이_잡혔는지_판단할_수_있다() {
         /*
 
@@ -485,75 +438,5 @@ public class BoardTest {
         softAssertions.assertThat(board.isKingCaptured()).isTrue();
 
         softAssertions.assertAll();
-    }
-
-    @Test
-    void 흰_진영의_기물_점수가_더_높다면_흰_진영이_더_유리하다() {
-        /*
-
-        .K......
-        ........
-        ........
-        ........
-        ........
-        ........
-        ........
-        ....rk..
-
-        */
-
-        boardMap.put(Position.of(E, ONE), new Rook(WHITE));
-        boardMap.put(Position.of(F, ONE), new King(WHITE));
-        boardMap.put(Position.of(B, EIGHT), new King(BLACK));
-        final Board board = new Board(boardMap);
-
-        assertThat(board.calculateAdvantageSide()).isEqualTo(WHITE);
-    }
-
-    @Test
-    void 검은_진영의_기물_점수가_더_높다면_검은_진영이_더_유리하다() {
-        /*
-
-        .KR.....
-        ........
-        ........
-        ........
-        ........
-        ........
-        ........
-        .....k..
-
-        */
-
-        boardMap.put(Position.of(F, ONE), new King(WHITE));
-        boardMap.put(Position.of(C, EIGHT), new Rook(BLACK));
-        boardMap.put(Position.of(B, EIGHT), new King(BLACK));
-        final Board board = new Board(boardMap);
-
-        assertThat(board.calculateAdvantageSide()).isEqualTo(BLACK);
-    }
-
-    @Test
-    void 두_진영의_기물_점수가_같다면_어떤_진영도_유리하지_않으므로_중립을_반환한다() {
-        /*
-
-        .KR.....
-        ........
-        ........
-        ........
-        ........
-        ........
-        ........
-        ....rk..
-
-        */
-
-        boardMap.put(Position.of(E, ONE), new Rook(WHITE));
-        boardMap.put(Position.of(F, ONE), new King(WHITE));
-        boardMap.put(Position.of(C, EIGHT), new Rook(BLACK));
-        boardMap.put(Position.of(B, EIGHT), new King(BLACK));
-        final Board board = new Board(boardMap);
-
-        assertThat(board.calculateAdvantageSide()).isEqualTo(NEUTRALITY);
     }
 }
